Add example code for gioui
For some reason, I could not run main.go on macOS from terminal. However, when I made a simple gioui gui app, it worked.

I am adding it as it's useful for quick testing and actually seeing how the dialog runs.

gioui is a simple UX framework that is multiplatform; what I like about it is you just need go with cgo
to build it, the app code is just go, and there are no other steps.

(Cross-building is still not really possible, though, as it is using cgo.)

I have made example_gioui a separate module so installing dialog doesn't have all gioui as a dependency.