Fix deprecated warnings for macOS
macOS has a new framework UniformTypeIdentifiers that needs to be used for this.

Note that the behavior is not exactly the same - if you add "jpg", it will find
a UTType that has "jpg" extension and will use that instead of creating a new type;
so it will match both "jpg" and "jpeg".

I think it is acceptable (and, probably, even better).